gallery/requirements_lock.txt
amaranth==0.5.8
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
jinja2==3.1.6
jschon==0.11.1
kiwisolver==1.4.8
markupsafe==3.0.3
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
py==1.11.0
greenlet==3.1.1
hypothesis==6.122.6
sortedcontainers==2.4.0
cffi==1.17.1
pycparser==2.22
attrs==25.1.0
exceptiongroup==1.2.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyvcd==0.4.1
pyyaml==6.0.2
rfc3986==2.0.0
six==1.17.0
